Rename GINT_STATIC_INTEGRALS cache variable and macro
Now the cache variable GINT_ENABLE_STATIC_INTEGRALS enables the use of
the static coulomb sturmian integral data. This sets / defines the
macro GINT_HAVE_STATIC_INTEGRALS in order to flag the availability of
the static integral data to the code.

This is done in order to clarify more what is going on and to be
consistent with the new optional gaussian libraries which will soon be
added.